DROID-3476 fixes
